fix!: Renamed action for marketplace
It was need for github market place

Your action.yml needs changes before it can be published.

curl-action - Name must be unique. Cannot match an existing action, user or organization name.

Add a sentence describing your Action in 125 characters or less.

For the record it was orginaly named curl-action that collide with:

https://github.com/marketplace/actions/curl-action

It looks like this one is not responding it can't be used.

https://github.com/marketplace/actions/curl

Maybe I'll find a better name someday.

Relate-to: https://github.com/rzr/curl-action/releases/new?marketplace=true
Signed-off-by: Philippe Coval <rzr@users.sf.net>
Change-Id: I3dd89c54c246ff5a31b055b1fad54cb56d3d753b